+++ lib/rley/tokens/token_range.rb
@@ -29,6 +29,8 @@ module Rley # This module is used as a namespace
             result = low == other[:low] && high == other[:high]
           when TokenRange
             result = low == other.low && high == other.high
+          when Range
+            result = low == other.first && high == other.last
           when Array
             result = low == other[0] && high == other[1]
         end